requirements.txt
opencv-python==4.8.0.76
mediapipe==0.10.3
Pillow==10.0.0
